\name{internal.utils}
\alias{extract.subsample}
\alias{extract.bootsample}
\alias{print.bootsample}
\alias{print.bootsample.rfsrc}
\alias{print.imbalanced.performance}
\alias{print.subsample}
\alias{print.subsample.rfsrc}
\alias{fast.save}
\alias{fast.load}
\alias{fast.save.list}
\alias{fast.load.list}
\alias{lsos}
\alias{get.imbalanced.performance}
\alias{get.imbalanced.optimize}
\alias{get.pr.auc}
\alias{get.pr.curve}
\alias{get.rfq.threshold}
\alias{get.auc}
\alias{get.auc.survival}
\alias{get.bayes.rule}
\alias{get.brier.error}
\alias{get.cindex}
\alias{get.confusion}
\alias{get.logloss}
\alias{get.misclass.error}
\alias{get.mv.cserror}
\alias{get.mv.csvimp}
\alias{get.mv.error}
\alias{get.mv.error.block}
\alias{get.mv.formula}
\alias{get.mv.predicted}
\alias{get.mv.vimp}
\alias{extract.quantile}
\alias{get.quantile}
\alias{get.quantile.crps}
\alias{get.quantile.stat}
\alias{sid.perf.metric}
\title{Internal Utility Functions}

\description{
  These are internal utility functions exported for advanced usage. This
  Rd file is used solely to register aliases.
}

\keyword{internal}
