\name{vdv}
\docType{data}
\alias{vdv}
\title{van de Vijver Microarray Breast Cancer}
\description{
  Gene expression profiling for predicting clinical outcome of breast
  cancer (van't Veer et al., 2002).  Microarray breast cancer data set
  of 4707 expression values on 78 patients with survival information.
}
\references{
  van't Veer L.J. et al. (2002).  Gene expression profiling predicts
  clinical outcome of breast cancer.  \emph{Nature}, \bold{12},
  530--536.
}
\examples{data(vdv, package = "randomForestSRC")}
\keyword{datasets}
