\name{veteran}
\docType{data}
\alias{veteran}
\title{Veteran's Administration Lung Cancer Trial}
\description{
  Randomized trial of two treatment regimens for lung cancer.
  This is a standard survival analysis data set.
}
\source{Kalbfleisch and Prentice, \emph{The Statistical
    Analysis of Failure Time Data.}}
\references{
  Kalbfleisch J. and Prentice R, (1980) \emph{The Statistical
    Analysis of Failure Time Data.}  New York: Wiley.
}
\examples{data(veteran, package = "randomForestSRC")}
\keyword{datasets}
